de l'action a tout va creer les bases dans exec=spip_listes si on est pas en 192 beta4
git-svn-id: svn://zone.spip.org/spip-zone/_plugins_@8972 ac52e18a-acf5-0310-9fe8-c4428f23b10a

+++ _test_/spip-listes/spip-listes_1_9_2/action/spiplistes_supprimer_abonne.php
@@ -2,27 +2,28 @@
 if (!defined("_ECRIRE_INC_VERSION")) return;
 
 include_spip('inc/extra_plus');
-function action_spiplistes_changer_statut_abonne_dist()
+include_spip('inc/autoriser');
+function action_spiplistes_supprimer_abonne_dist()
 {
 	$securiser_action = charger_fonction('securiser_action', 'inc');
 	$id_auteur = $securiser_action();
 	$redirect = urldecode(_request('redirect'));
 
 	//changer de statut
-	//if (autoriser())
-	$result = spip_query("SELECT * FROM spip_auteurs WHERE id_auteur="._q($id_auteur));
-	if ($row = spip_fetch_array($result)) {
-		$id_auteur=$row['id_auteur'];
-		$statut=$row['statut'];
-
-		if($statut=='6forum'){
-			spip_query("DELETE FROM spip_abonnes_listes WHERE id_auteur="._q($id_auteur));
-			spip_query("DELETE FROM spip_auteurs WHERE id_auteur="._q($id_auteur));
-		}
-	}
+	if (autoriser('supprimer','auteur',$id_auteur)){
+		$result = spip_query("SELECT * FROM spip_auteurs WHERE id_auteur="._q($id_auteur));
+		if ($row = spip_fetch_array($result)) {
+			$id_auteur=$row['id_auteur'];
+			$statut=$row['statut'];
 	
+			if($statut=='6forum'){
+				spip_query("DELETE FROM spip_abonnes_listes WHERE id_auteur="._q($id_auteur));
+				spip_query("DELETE FROM spip_auteurs WHERE id_auteur="._q($id_auteur));
+			}
+		}
+	}	
 	if ($redirect){
-		redirige_par_entete(str_replace("&amp;","&",$redirect);
+		redirige_par_entete(str_replace("&amp;","&",$redirect));
 	}
 }
